+++ atramhasis/scripts/generate_ldf_config.py
@@ -93,10 +93,12 @@ is set to the atramhasis.ldf.config_location from your ini file.'
     if ldf_protocol:
         ldfconfig['protocol'] = ldf_protocol
 
+    pids = []
     for p in skos_registry.get_providers():
         if any([not_shown in p.get_metadata()['subject'] for not_shown in ['external']]):
             continue;
         pid = p.get_metadata()['id']
+        pids.append(pid)
         filename = os.path.join(dump_location, '%s-full' % pid)
         filename_ttl = filename + '.ttl'
         filename_hdt = filename + '.hdt'
@@ -118,6 +120,17 @@ is set to the atramhasis.ldf.config_location from your ini file.'
                 sourceconfig['description'] = n.note
                 break
         ldfconfig['datasources'][pid] = sourceconfig
+
+    if len(pids):
+        sourceconfig = {
+            'title': 'All conceptschemes',
+            'type': 'CompositeDatasource',
+            'description': 'All conceptschemes contained in this Atramhasis instance together.',
+            'settings': {
+                'references': pids
+            }
+        }
+        ldfconfig['datasources']['composite'] = sourceconfig
 
 
     config_filename = os.path.join(config_location, 'ldf_server_config.json')